Update CloudAdapter and Managed Identity in bot-proactive-messaging-teamsfx Sample
