--- samples/bot-proactive-messaging-teamsfx/.gitattributes
@@ -1,2 +1,0 @@
-# Auto detect text files and perform LF normalization
-* text=auto
