+++ samples/bot-proactive-messaging-teamsfx/.gitignore
@@ -1,15 +1,19 @@
-node_modules
-.fx/configs/localSettings.json
-.fx/states/*.userdata
-.DS_Store
-.archive
-.archive.log
-.env.teamsfx.local
-subscriptionInfo.json
-build
-.fx/configs/config.local.json
-.fx/states/state.local.json
-.backup
+# TeamsFx files
 env/.env.*.user
-.backup/*
+env/.env.local
+.localConfigs.testTool
+.localConfigs
+.notification.localstore.json
+.notification.testtoolstore.json
+appPackage/build
+
+# dependencies
+node_modules/
+
+# misc
+.env
+.deployment
+.DS_Store
+
+# Dev tool directories
 /devTools/
