+++ samples/bot-proactive-messaging-teamsfx/.vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+  "recommendations": [
+    "TeamsDevApp.ms-teams-vscode-extension"
+  ]
+}
