+++ samples/bot-proactive-messaging-teamsfx/.webappignore
@@ -1,0 +1,28 @@
+.webappignore
+.fx
+.deployment
+.localConfigs.testTool
+.localConfigs
+.notification.localstore.json
+.notification.testtoolstore.json
+.vscode
+*.js.map
+*.ts.map
+*.ts
+.git*
+.tsbuildinfo
+CHANGELOG.md
+readme.md
+local.settings.json
+test
+tsconfig.json
+.DS_Store
+teamsapp.yml
+teamsapp.*.yml
+/env/
+/node_modules/.bin
+/node_modules/ts-node
+/node_modules/typescript
+/appPackage/
+/infra/
+/devTools/
