+++ samples/bot-proactive-messaging-teamsfx/appPackage/manifest.json
@@ -1,5 +1,5 @@
 {
-    "$schema": "https://developer.microsoft.com/json-schemas/teams/v1.19/MicrosoftTeams.schema.json",
+    "$schema": "https://developer.microsoft.com/en-us/json-schemas/teams/v1.19/MicrosoftTeams.schema.json",
     "manifestVersion": "1.19",
     "version": "1.0.0",
     "id": "${{TEAMS_APP_ID}}",
@@ -7,20 +7,20 @@
         "name": "Teams App, Inc.",
         "websiteUrl": "https://www.example.com",
         "privacyUrl": "https://www.example.com/privacy",
-        "termsOfUseUrl": "https://www.example.com/termsofuse"
+        "termsOfUseUrl": "https://www.example.com/termofuse"
     },
     "icons": {
-        "color": "resources/color.png",
-        "outline": "resources/outline.png"
+        "color": "color.png",
+        "outline": "outline.png"
     },
     "name": {
-        "short": "${{CONFIG_APPNAME_SHORT}}",
-        "full": "${{CONFIG_MANIFEST_APPNAME_FULL}}"
+        "short": "Proactive Messaging",
+        "full": "This bot sends proactive reminders by saving conversation references in Teams."
     },
-  "description": {
-    "short": "Bot sample for proactive notifications ${{CONFIG_APPNAME_SHORT}}",
-    "full": "This bot sends proactive reminders by saving conversation references in Teams. ${{CONFIG_APPNAME_SHORT}}"
-  },
+    "description": {
+        "short": "This bot sends proactive reminders by saving conversation references in Teams.",
+        "full": "This bot sample demonstrates proactive messaging in Microsoft Teams by storing conversation references for sending reminder notifications. It includes guidance for local debugging, deployment to Azure, and integration with Teams Toolkit."
+    },
     "accentColor": "#FFFFFF",
     "bots": [
         {
@@ -34,21 +34,17 @@
             "isNotificationOnly": false,
             "commandLists": [
                 {
-                    "scopes": [
-                        "personal",
-                        "team",
-                        "groupChat"
-                    ],
-                    "commands": [
-                        {
-                            "title": "welcome",
-                            "description": "Resend welcome card of this Bot"
-                        },
-                        {
-                            "title": "learn",
-                            "description": "Learn about Adaptive Card and Bot Command"
-                        }
-                    ]
+                  "scopes": ["personal", "team", "groupChat"],
+                  "commands": [
+                    {
+                        "title": "welcome",
+                        "description": "Resend welcome card of this Bot"
+                    },
+                    {
+                        "title": "learn",
+                        "description": "Learn about Adaptive Card and Bot Command"
+                    }
+                  ]
                 }
             ]
         }
@@ -60,7 +56,5 @@
         "identity",
         "messageTeamMembers"
     ],
-    "validDomains": [
-        "${{PROVISIONOUTPUT__BOTOUTPUT__VALIDDOMAIN}}"
-    ]
+    "validDomains": ["${{BOT_DOMAIN}}"]
 }
