--- samples/bot-proactive-messaging-teamsfx/bot/.gitignore
@@ -1,7 +1,0 @@
-# dependencies
-node_modules/
-
-# misc
-.env
-.deployment
-.DS_Store
