--- samples/bot-proactive-messaging-teamsfx/bot/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "teams-conversation-bot",
-  "version": "1.0.0",
-  "msteams": {
-    "teamsAppId": null
-  },
-  "description": "Microsoft Teams Toolkit hello world Bot sample",
-  "engines": {
-    "node": "14 || 16 || 18"
-  },
-  "author": "Microsoft",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev:teamsfx": "env-cmd --silent -f .env.teamsfx.local npm run dev",
-    "dev": "nodemon --inspect=9239 --signal SIGINT ./index.js",
-    "start": "node ./index.js",
-    "watch": "nodemon ./index.js"
-  },
-  "dependencies": {
-    "@microsoft/adaptivecards-tools": "^0.1.2",
-    "botbuilder": "~4.15.0",
-    "botbuilder-dialogs": "~4.5.0",
-    "isomorphic-fetch": "^3.0.0",
-    "dotenv": "^16.3.1",
-    "restify": "^10.0.0"
-  },
-  "devDependencies": {
-    "env-cmd": "^10.1.0",
-    "nodemon": "^2.0.7"
-  }
-}
