+++ samples/bot-proactive-messaging-teamsfx/config.js
@@ -1,0 +1,8 @@
+const config = {
+  MicrosoftAppId: process.env.BOT_ID,
+  MicrosoftAppType: process.env.BOT_TYPE,
+  MicrosoftAppTenantId: process.env.BOT_TENANT_ID,
+  MicrosoftAppPassword: process.env.BOT_PASSWORD,
+};
+
+module.exports = config;
