+++ samples/bot-proactive-messaging-teamsfx/package.json
@@ -1,23 +1,33 @@
 {
-    "name": "ProactiveMessagesTeamsFx",
-    "version": "0.0.1",
-    "description": "",
-    "author": "",
+    "name": "proactivemessagesteamsfx",
+    "version": "1.0.0",
+    "msteams": {
+      "teamsAppId": null
+    },
+    "description": "Microsoft Teams Toolkit hello world Bot sample",
+    "engines": {
+        "node": "18 || 20"
+    },
+    "author": "Microsoft",
+    "license": "MIT",
+    "main": "index.js",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "install:bot": "cd bot && npm install",
-        "installAll": "npm run install:bot",
-        "build": "node build.js"
+        "dev:teamsfx": "env-cmd --silent -f .localConfigs npm run dev",
+        "dev:teamsfx:testtool": "env-cmd --silent -f .localConfigs.testTool npm run dev",
+        "dev:teamsfx:launch-testtool": "env-cmd --silent -f env/.env.testtool teamsapptester start",
+        "dev": "nodemon --inspect=9239 --signal SIGINT ./index.js",
+        "start": "node ./index.js",
+        "watch": "nodemon ./index.js",
+        "test": "echo \"Error: no test specified\" && exit 1"
     },
     "dependencies": {
-        "botbuilder": "^4.18.0",
-        "botbuilder-dialogs": "^4.18.0",
+        "botbuilder": "^4.23.1",
         "dotenv": "^16.3.1",
-        "restify": "^11.1.0"
+        "express": "^5.0.1"
     },
     "devDependencies": {
-        "@microsoft/teamsfx-cli": "0.*",
+        "env-cmd": "^10.1.0",
+        "nodemon": "^3.1.7",
         "esbuild": "^0.18.1"
-    },
-    "license": "MIT"
+    }
 }
