--- samples/bot-proactive-messaging-teamsfx/permissions.json
@@ -1,9 +1,0 @@
-[
-    {
-        "resource": "Microsoft Graph",
-        "delegated": [
-            "User.Read"
-        ],
-        "application": []
-    }
-]
