--- samples/bot-proactive-messaging-teamsfx/templates/appManifest/aad.template.json
@@ -1,101 +1,0 @@
-{
-    "id": "{{state.fx-resource-aad-app-for-teams.objectId}}",
-    "appId": "{{state.fx-resource-aad-app-for-teams.clientId}}",
-    "name": "{{config.manifest.appName.short}}-aad",
-    "accessTokenAcceptedVersion": 2,
-    "signInAudience": "AzureADMyOrg",
-    "optionalClaims": {
-        "idToken": [],
-        "accessToken": [
-            {
-                "name": "idtyp",
-                "source": null,
-                "essential": false,
-                "additionalProperties": []
-            }
-        ],
-        "saml2Token": []
-    },
-    "requiredResourceAccess": [
-        {
-            "resourceAppId": "Microsoft Graph",
-            "resourceAccess": [
-                {
-                    "id": "User.Read",
-                    "type": "Scope"
-                }
-            ]
-        }
-    ],
-    "oauth2Permissions": [
-        {
-            "adminConsentDescription": "Allows Teams to call the app's web APIs as the current user.",
-            "adminConsentDisplayName": "Teams can access app's web APIs",
-            "id": "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}",
-            "isEnabled": true,
-            "type": "User",
-            "userConsentDescription": "Enable Teams to call this app's web APIs with the same rights that you have",
-            "userConsentDisplayName": "Teams can access app's web APIs and make requests on your behalf",
-            "value": "access_as_user"
-        }
-    ],
-    "preAuthorizedApplications": [
-        {
-            "appId": "1fec8e78-bce4-4aaf-ab1b-5451cc387264",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        },
-        {
-            "appId": "5e3ce6c0-2b1f-4285-8d4b-75ee78787346",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        },
-        {
-            "appId": "d3590ed6-52b3-4102-aeff-aad2292ab01c",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        },
-        {
-            "appId": "00000002-0000-0ff1-ce00-000000000000",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        },
-        {
-            "appId": "bc59ab01-8403-45c6-8796-ac3ef710b3e3",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        },
-        {
-            "appId": "0ec893e0-5785-4de6-99da-4ed124e5296c",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        },
-        {
-            "appId": "4765445b-32c6-49b0-83e6-1d93765276ca",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        },
-        {
-            "appId": "4345a7b9-9a63-4910-a426-35363201d503",
-            "permissionIds": [
-                "{{state.fx-resource-aad-app-for-teams.oauth2PermissionScopeId}}"
-            ]
-        }
-    ],
-    "identifierUris": [
-        "{{state.fx-resource-aad-app-for-teams.applicationIdUris}}"
-    ],
-    "replyUrlsWithType": [
-        {
-            "url": "{{state.fx-resource-aad-app-for-teams.botEndpoint}}/auth-end.html",
-            "type": "Web"
-        }
-    ]
-}
