--- samples/bot-proactive-messaging-teamsfx/templates/azure/azure.parameters.dev.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#",
-  "contentVersion": "1.0.0.0",
-  "parameters": {
-    "provisionParameters": {
-      "value": {
-        "resourceBaseName": "proactivemdev${{RESOURCE_SUFFIX}}",
-        "botAadAppClientId": "${{BOT_ID}}",
-        "botAadAppClientSecret": "${{SECRET_BOT_PASSWORD}}"
-      }
-    }
-  }
-}
