--- samples/bot-proactive-messaging-teamsfx/templates/azure/config.bicep
@@ -1,15 +1,0 @@
-@secure()
-param provisionParameters object
-param provisionOutputs object
-// Get existing app settings for merge
-var botCurrentAppSettings = list('${provisionOutputs.botOutput.value.botWebAppResourceId}/config/appsettings', '2021-02-01').properties
-
-// Merge TeamsFx configurations to Bot resources
-module teamsFxBotConfig './teamsFx/bot.bicep' = {
-  name: 'addTeamsFxBotConfiguration'
-  params: {
-    provisionParameters: provisionParameters
-    provisionOutputs: provisionOutputs
-    currentAppSettings: botCurrentAppSettings
-  }
-}
