--- samples/bot-proactive-messaging-teamsfx/templates/azure/main.bicep
@@ -1,18 +1,0 @@
-@secure()
-param provisionParameters object
-
-module provision './provision.bicep' = {
-  name: 'provisionResources'
-  params: {
-    provisionParameters: provisionParameters
-  }
-}
-output provisionOutput object = provision
-module teamsFxConfig './config.bicep' = {
-  name: 'addTeamsFxConfigurations'
-  params: {
-    provisionParameters: provisionParameters
-    provisionOutputs: provision
-  }
-}
-output teamsFxConfigurationOutput object = contains(reference(resourceId('Microsoft.Resources/deployments', teamsFxConfig.name), '2020-06-01'), 'outputs') ? teamsFxConfig : {}
