--- samples/bot-proactive-messaging-teamsfx/templates/azure/teamsFx/bot.bicep
@@ -1,24 +1,0 @@
-// Auto generated content, please customize files under provision folder
-
-@secure()
-param provisionParameters object
-param provisionOutputs object
-@secure()
-param currentAppSettings object
-
-var botWebAppName = split(provisionOutputs.botOutput.value.botWebAppResourceId, '/')[8]
-
-var botAadAppClientSecret = provisionParameters['botAadAppClientSecret']
-
-var botId = provisionParameters['botAadAppClientId']
-
-resource botWebAppSettings 'Microsoft.Web/sites/config@2021-02-01' = {
-  name: '${botWebAppName}/appsettings'
-  properties: union({
-    INITIATE_LOGIN_ENDPOINT: uri(provisionOutputs.botOutput.value.siteEndpoint, 'auth-start.html') // The page is used to let users consent required OAuth permissions during bot SSO process
-    BOT_ID: botId // ID of your bot
-    BOT_PASSWORD: botAadAppClientSecret // Secret of your bot
-    IDENTITY_ID: provisionOutputs.identityOutput.value.identityClientId // User assigned identity id, the identity is used to access other Azure resources
-    PROVISIONOUTPUT_BOTOUTPUT_SITEENDPOINT : provisionOutputs.botOutput.value.siteEndpoint // Site endpoint of AAD application
-  }, currentAppSettings)
-}
